requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
flake8==7.1.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.2
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing-extensions==4.12.2
tzdata==2024.1
